requirements.txt
torch
ftfy
regex
tqdm
Pillow
clip-by-openai
openai
streamlit
matplotlib
